pyproject.toml
[project]
name = "hackathon-demo"
version = "0.1.0"
description = "Demo project for Agentic AI Hackathon."
readme = "README.md"
authors = [
    { name = "laiba", email = "laiba@retrocausal.ai" }
]
requires-python = ">=3.12"
dependencies = [
    "chainlit>=2.8.1",
    "openai-agents>=0.3.2",
    "python-dotenv>=1.1.1",
]

[project.scripts]
hackathon-demo = "hackathon_demo:main"
